view api med_store

+++ medical_store/models.py
@@ -22,11 +22,10 @@ class MedicineInventory(models.Model):
 
 class CompanyDetails(models.Model):
     company_name = models.CharField(max_length=200)
-    company_contact = models.PositiveIntegerField(validators=[MaxValueValidator(9999999999),MinValueValidator(1000000000)])
+    company_contact = models.BigIntegerField(validators=[MaxValueValidator(9999999999),MinValueValidator(1000000000)])
     company_address=models.CharField(max_length=200)
     company_email = models.EmailField()
     gst_number = models.CharField(max_length=15)
-   # account = models.ForeignKey(User,on_delete=models.CASCADE)
 
     def __str__(self):
         return self.company_name
